third_party/WebKit/LayoutTests/fast/mediastream/MediaStreamConstructor-expected.txt

Tests MediaStream constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Interface tests.
PASS typeof MediaStream is 'function'
PASS MediaStream.length is 0
PASS MediaStream.name is 'MediaStream'
PASS Object.getPrototypeOf(MediaStream.prototype) is EventTarget.prototype
PASS window.MediaStream is window.webkitMediaStream
Got local stream.
PASS localStream.getAudioTracks().length is 1
PASS localStream.getVideoTracks().length is 1
PASS checkIdAttribute(localStream.id) is true
PASS new MediaStream(document) threw exception TypeError: Failed to construct 'MediaStream': No matching constructor signature..
PASS new MediaStream([document]) threw exception TypeError: Failed to construct 'MediaStream': Failed to convert value to 'MediaStreamTrack'..
PASS new MediaStream([stream.getAudioTracks()[0], document]) threw exception TypeError: Failed to construct 'MediaStream': Failed to convert value to 'MediaStreamTrack'..
PASS new MediaStream([null]) threw exception TypeError: Failed to construct 'MediaStream': Failed to convert value to 'MediaStreamTrack'..
PASS new MediaStream([undefined]) threw exception TypeError: Failed to construct 'MediaStream': Failed to convert value to 'MediaStreamTrack'..
PASS new MediaStream(null) threw exception TypeError: Failed to construct 'MediaStream': No matching constructor signature..
PASS new MediaStream(undefined) threw exception TypeError: Failed to construct 'MediaStream': No matching constructor signature..
Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS newStream.active is false
PASS checkIdAttribute(newStream.id) is true
Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS newStream.active is false
PASS checkIdAttribute(newStream.id) is true
Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS newStream.active is true
PASS checkIdAttribute(newStream.id) is true
Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS newStream.active is true
PASS checkIdAttribute(newStream.id) is true
Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS newStream.active is true
PASS checkIdAttribute(newStream.id) is true
Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS newStream.active is true
PASS checkIdAttribute(newStream.id) is true
Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS newStream.active is true
PASS checkIdAttribute(newStream.id) is true
Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS newStream.active is false
PASS checkIdAttribute(newStream.id) is true
Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS newStream.active is false
PASS checkIdAttribute(newStream.id) is true
Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS newStream.active is false
PASS checkIdAttribute(newStream.id) is true
Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS newStream.active is false
PASS checkIdAttribute(newStream.id) is true
PASS successfullyParsed is true

TEST COMPLETE
